UC-4193 StateQueueClient 100ms hard coded timeout - refactoring before actual fix: - BlockingTcpClient moved to separate source/include files - friendship of StateQueueClient and BlockingTcpClient removed to avoid interface confusion - some of the BlockingTcpClient methods made private
